refactor to remove code duplication
Refactor to remove code duplication.  Move all functions under one
folder and reference functions with a cloudformation `Handler`
parameter.  Also update tests accordingly.
