--- set_instance_tags/requirements.txt
@@ -1,2 +1,0 @@
-crhelper~=2.0.6
-synapseclient~=2.2.0
