+++ template.yaml
@@ -18,7 +18,8 @@ Resources:
   SetBucketTagsFunction:
     Type: AWS::Serverless::Function
     Properties:
-      CodeUri: set_bucket_tags/
+      CodeUri: set_tags/
+      Handler: set_bucket_tags.handler
       Role: !GetAtt SetBucketTagsFunctionRole.Arn
       Environment:
         Variables:
@@ -57,7 +58,8 @@ Resources:
   SetInstanceTagsFunction:
     Type: AWS::Serverless::Function
     Properties:
-      CodeUri: set_instance_tags/
+      CodeUri: set_tags/
+      Handler: set_instance_tags.handler
       Role: !GetAtt SetInstanceTagsFunctionRole.Arn
       Environment:
         Variables:
