--- tests/unit/set_instance_tags/test_get_principal_id.py
@@ -1,24 +1,0 @@
-import unittest
-
-from set_instance_tags import app
-
-
-class TestGetPrincipalId(unittest.TestCase):
-
-  def test_tag_present(self):
-    tags = [
-      {'Key': 'heresatag', 'Value': 'heresatagvalue'},
-      {'Key': 'theresatag', 'Value': 'theresatagvalue'},
-      {'Key': 'aws:servicecatalog:provisioningPrincipalArn', 'Value': 'foo/bar'}
-    ]
-    result = app.get_principal_id(tags)
-    self.assertEqual(result, 'bar')
-
-
-  def test_tag_missing(self):
-    with self.assertRaises(ValueError):
-      tags = [
-        {'Key': 'heresatag', 'Value': 'heresatagvalue'},
-        {'Key': 'theresatag', 'Value': 'theresatagvalue'}
-      ]
-      app.get_principal_id(tags)
