+++ tests/unit/set_tags/test_get_bucket_name.py
@@ -1,6 +1,6 @@
 import unittest
 
-from set_bucket_tags import app
+from set_tags import set_bucket_tags
 
 
 class TestGetBucketName(unittest.TestCase):
@@ -11,11 +11,11 @@ class TestGetBucketName(unittest.TestCase):
         'BucketName': 'a_bucket_name'
       }
     }
-    result = app.get_bucket_name(event)
+    result = set_bucket_tags.get_bucket_name(event)
     self.assertEqual(result, 'a_bucket_name')
 
 
   def test_bucketname_missing(self):
     with self.assertRaises(ValueError):
       event = { 'ResourceProperties': {} }
-      app.get_bucket_name(event)
+      set_bucket_tags.get_bucket_name(event)
