+++ tests/unit/set_tags/test_get_instance_id.py
@@ -1,6 +1,6 @@
 import unittest
 
-from set_instance_tags import app
+from set_tags import set_instance_tags
 
 
 class TestGetInstanceId(unittest.TestCase):
@@ -11,11 +11,11 @@ class TestGetInstanceId(unittest.TestCase):
         'InstanceId': 'a_instance_id'
       }
     }
-    result = app.get_instance_id(event)
+    result = set_instance_tags.get_instance_id(event)
     self.assertEqual(result, 'a_instance_id')
 
 
   def test_instance_id_missing(self):
     with self.assertRaises(ValueError):
       event = { 'ResourceProperties': {} }
-      app.get_instance_id(event)
+      set_instance_tags.get_instance_id(event)
