+++ tests/unit/utils/test_get_synapse_user_profile.py
@@ -2,7 +2,7 @@ import unittest
 
 from synapseclient.core.exceptions import SynapseHTTPError
 from unittest.mock import patch, MagicMock
-from set_bucket_tags import app
+from set_tags import utils
 
 TEST_USER_PROFILE = {
   "createdOn": "2020-06-18T16:34:18.000Z",
@@ -26,11 +26,11 @@ class TestGetSynapseUserProfile(unittest.TestCase):
   @patch('synapseclient.Synapse')
   def test_valid_id(self, MockSynapse):
     MockSynapse.return_value.getUserProfile=mock_get_user_profile
-    result = app.get_synapse_user_profile("1111111")
+    result = utils.get_synapse_user_profile("1111111")
     self.assertEqual(result, TEST_USER_PROFILE)
 
   @patch('synapseclient.Synapse')
   def test_invalid_id(self, MockSynapse):
     MockSynapse.return_value.getUserProfile=mock_get_user_profile
     with self.assertRaises(SynapseHTTPError):
-        app.get_synapse_user_profile("3333333")
+      utils.get_synapse_user_profile("3333333")
