+++ tests/unit/utils/test_get_synapse_user_profile_tags.py
@@ -1,6 +1,6 @@
 import unittest
 
-from set_bucket_tags import app
+from set_tags import utils
 
 TEST_USER_PROFILE = {
   "createdOn": "2020-06-18T16:34:18.000Z",
@@ -14,7 +14,7 @@ TEST_USER_PROFILE = {
 class TestGetSynapseUserProfileTags(unittest.TestCase):
 
   def test_happy_default_ignore(self):
-    result = app.get_synapse_user_profile_tags(TEST_USER_PROFILE)
+    result = utils.get_synapse_user_profile_tags(TEST_USER_PROFILE)
     expected = [
         {'Key': 'synapse:firstName', 'Value': 'Joe'},
         {'Key': 'synapse:lastName', 'Value': 'Smith'},
@@ -27,7 +27,7 @@ class TestGetSynapseUserProfileTags(unittest.TestCase):
     self.assertListEqual(result, expected)
 
   def test_happy_mulitple_ignores(self):
-    result = app.get_synapse_user_profile_tags(TEST_USER_PROFILE,
+    result = utils.get_synapse_user_profile_tags(TEST_USER_PROFILE,
                                                ["createdOn","company","firstName","lastName"])
     expected = [
         {'Key': 'synapse:ownerId', 'Value': '1111111'},
@@ -38,7 +38,7 @@ class TestGetSynapseUserProfileTags(unittest.TestCase):
     self.assertListEqual(result, expected)
 
   def test_happy_ignores_user_name(self):
-    result = app.get_synapse_user_profile_tags(TEST_USER_PROFILE, ["userName"])
+    result = utils.get_synapse_user_profile_tags(TEST_USER_PROFILE, ["userName"])
     expected = [
         {'Key': 'synapse:createdOn', 'Value': '2020-06-18T16:34:18.000Z'},
         {'Key': 'synapse:firstName', 'Value': 'Joe'},
