+++ tests/unit/utils/test_get_synapse_user_team_id.py
@@ -1,9 +1,8 @@
 import unittest
 import boto3
-import synapseclient
 
 from unittest.mock import patch
-from set_bucket_tags import app
+from set_tags import utils
 from botocore.stub import Stubber
 
 
@@ -20,10 +19,10 @@ class TestGetSynapseUserTeamId(unittest.TestCase):
     ssm = boto3.client('ssm')
     with Stubber(ssm) as stubber, \
       patch('synapseclient.Synapse') as syn_mock, \
-      patch('set_bucket_tags.app.get_synapse_team_ids') as team_ids_mock:
+      patch('set_tags.utils.get_synapse_team_ids') as team_ids_mock:
         syn_mock.return_value.get_membership_status = get_membership_status_is_member
         team_ids_mock.return_value = ["1111111","2222222"]
-        result = app.get_synapse_user_team_id(1111111, ["1111111","2222222"])
+        result = utils.get_synapse_user_team_id(1111111, ["1111111","2222222"])
         expected = "1111111"
         self.assertEqual(result, expected)
 
@@ -32,9 +31,9 @@ class TestGetSynapseUserTeamId(unittest.TestCase):
     ssm = boto3.client('ssm')
     with Stubber(ssm) as stubber, \
       patch('synapseclient.Synapse') as syn_mock, \
-      patch('set_bucket_tags.app.get_synapse_team_ids') as team_ids_mock:
+      patch('set_tags.utils.get_synapse_team_ids') as team_ids_mock:
         syn_mock.return_value.get_membership_status = get_membership_status_is_not_member
         team_ids_mock.return_value = ["1111111","2222222"]
-        result = app.get_synapse_user_team_id(3333333, ["1111111","2222222"])
+        result = utils.get_synapse_user_team_id(3333333, ["1111111","2222222"])
         expected = None
         self.assertEqual(result, expected)
